api/requirements.txt
django
djangorestframework
django-types
django-polymorphic
